*Allowed the Editor to display Musashi, Hanzou and Jogurt's Promoted Stats
In all my testing, modifying their promoted stats didn't break anything in the game. It's possible that those spots were just filled with Junk data sometime during the game's development. If issues arise with this feature, I'll revert and give the option to show their promoted stats, and do a repoint for them.